Revert "second timmer"
